archive/python_model/requirements.txt
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
fastapi==0.104.0
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pydantic==2.5.0
numpy==1.24.0
